components file refactoring
